Refactor some tensor tests
